Clean up
Based on the statement from others, The code was cleaned, variable are written in the right forma, unnecessary modifiers have been cleaned and the contract runs optimally. As far as function ality goes, this is meant to mimic a wallet and wallet are required to be simple and fast, i have tackled the issue of simple by making it possible to send funds using username as most wallet are now doing, i planned to make other features but that would only be possible using a node. Thank you for viewing this submission